requirements.txt
python-dotenv~=1.1.1
google-api-client-wrapper>=2.0.0
langgraph~=0.6.8
langchain-core~=0.3.78
langchain-google-genai~=2.1.12
langchain~=0.3.27
textwrap3
pydantic~=2.12.0
python-telegram-bot>=21.0

python-telegram-bot
flask
google-api-python-client~=2.163.0
google-auth-httplib2
google-auth-oauthlib~=1.2.1